feat: NearDeterministicAccount (#42)
Add a new kind of Near implicit account, which has the initial state and
WASM code encoded in its id.

This is the main feature of NEP-616.
https://github.com/near/NEPs/pull/616

I propose to keep the actual derivation of the id outside of this crate.
The derivation can live in near-primitives alongside other implicit id
derivation functions.
See
https://github.com/near/nearcore/blob/90c8b2504a8b2333082f4f937f3d581c2e4ec329/core/primitives/src/utils.rs#L446-L469.